+++ bot/lib/tasks/__init__.py
@@ -61,7 +61,8 @@ class Loop:
         self._next_iteration = None
 
         if not inspect.iscoroutinefunction(self.coro):
-            raise TypeError('Expected coroutine function, not {0.__name__!r}.'.format(type(self.coro)))
+            raise TypeError(
+                'Expected coroutine function, not {0.__name__!r}.'.format(type(self.coro)))
 
     async def _call_loop_function(self, name, *args, **kwargs):
         coro = getattr(self, '_' + name)
@@ -167,7 +168,8 @@ class Loop:
         """
 
         if self._task is not None and not self._task.done():
-            raise RuntimeError('Task is already launched and is not completed.')
+            raise RuntimeError(
+                'Task is already launched and is not completed.')
 
         if self._injected is not None:
             args = (self._injected, *args)
@@ -253,7 +255,8 @@ class Loop:
             if not inspect.isclass(exc):
                 raise TypeError('{0!r} must be a class.'.format(exc))
             if not issubclass(exc, BaseException):
-                raise TypeError('{0!r} must inherit from BaseException.'.format(exc))
+                raise TypeError(
+                    '{0!r} must inherit from BaseException.'.format(exc))
 
         self._valid_exception = (*self._valid_exception, *exceptions)
 
@@ -280,7 +283,8 @@ class Loop:
             Whether all exceptions were successfully removed.
         """
         old_length = len(self._valid_exception)
-        self._valid_exception = tuple(x for x in self._valid_exception if x not in exceptions)
+        self._valid_exception = tuple(
+            x for x in self._valid_exception if x not in exceptions)
         return len(self._valid_exception) == old_length - len(exceptions)
 
     def get_task(self):
@@ -307,8 +311,10 @@ class Loop:
 
     async def _error(self, *args):
         exception = args[-1]
-        print('Unhandled exception in internal background task {0.__name__!r}.'.format(self.coro), file=sys.stderr)
-        traceback.print_exception(type(exception), exception, exception.__traceback__, file=sys.stderr)
+        print('Unhandled exception in internal background task {0.__name__!r}.'.format(
+            self.coro), file=sys.stderr)
+        traceback.print_exception(
+            type(exception), exception, exception.__traceback__, file=sys.stderr)
 
     def before_loop(self, coro):
         """A decorator that registers a coroutine to be called before the loop starts running.