+++ bot/modules/loader.py
@@ -1,7 +1,8 @@
-main_cogs = ["cogs.admin"]
-standard_cogs = ["cogs.register", "cogs.matches", "cogs.lobby"]
-__isGlobalLocked = True  # Lock the bot from getting messages
+from discord.ext.commands.errors import ExtensionAlreadyLoaded
 
+main_cogs = ["cogs.admin"]
+standard_cogs = ["cogs.register", "cogs.matches", "cogs.lobby", "cogs.muted"]
+__isGlobalLocked = True  # Lock the bot from getting messages
 
 def init(client):
     for cog in main_cogs:
@@ -17,7 +18,10 @@ def lockAll(client):
 
 def unlockAll(client):
     for cog in standard_cogs:
-        client.load_extension(cog)
+        try:
+            client.load_extension(cog)
+        except ExtensionAlreadyLoaded:
+            pass
     global __isGlobalLocked
     __isGlobalLocked = False